@@ -1,113 +1,100 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<TargetFrameworks>net7.0-ios;net7.0-android;net7.0-maccatalyst</TargetFrameworks>
-		<SingleProject>true</SingleProject>
-		<OutputType>Exe</OutputType>
-		<!-- Display name -->
-		<ApplicationTitle>ProjectSBS</ApplicationTitle>
-		<!-- App Identifier -->
-		<ApplicationId>com.morning4coffe.ProjectSBS</ApplicationId>
-		<ApplicationIdGuid>9B557464-9008-4AD5-8293-3A1CAD721DD7</ApplicationIdGuid>
-		<!-- Versions -->
-		<ApplicationDisplayVersion>1.0</ApplicationDisplayVersion>
-		<ApplicationVersion>1</ApplicationVersion>
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>net8.0-android;net8.0-ios;net8.0-maccatalyst</TargetFrameworks>
+    <TargetFrameworks Condition="'$(OverrideTargetFrameworks)'!=''">$(OverrideTargetFrameworks)</TargetFrameworks>
+    <SingleProject>true</SingleProject>
+    <OutputType>Exe</OutputType>
+    <!-- Display name -->
+    <ApplicationTitle>ProjectSBS</ApplicationTitle>
+    <!-- App Identifier -->
+    <ApplicationId>com.morning4coffe.ProjectSBS</ApplicationId>
+    <ApplicationIdGuid>01E54658-2C2B-44EA-907E-AD8EC550FE89</ApplicationIdGuid>
+    <!-- Versions -->
+    <ApplicationDisplayVersion>1.0</ApplicationDisplayVersion>
+    <ApplicationVersion>1</ApplicationVersion>
 
-		<AndroidManifest>Android\AndroidManifest.xml</AndroidManifest>
+    <AndroidManifest>Android\AndroidManifest.xml</AndroidManifest>
 
-		<!-- Debugger workaround https://github.com/dotnet/maui-samples/blob/8aa6b8780b12e97b157514c3bdc54bb4a13001cd/HelloMacCatalyst/HelloMacCatalyst.csproj#L7 -->
-		<!-- <MtouchExtraArgs Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'maccatalyst'">$(MtouchExtraArgs) -setenv:MONO_THREADS_SUSPEND=preemptive</MtouchExtraArgs> -->
-		<!-- Required for C# Hot Reload -->
-		<UseInterpreter Condition="'$(Configuration)' == 'Debug' and $([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) != 'maccatalyst'">True</UseInterpreter>
-		<IsUnoHead>true</IsUnoHead>
-	</PropertyGroup>
-	
-	<!--TODO Issue with AOT on .NET 7 with Graph and Toolkit-->
-	<PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|net7.0-android|AnyCPU'">
-		<RunAOTCompilation>False</RunAOTCompilation>
-		<AndroidEnableProfiledAot>False</AndroidEnableProfiledAot>
-	</PropertyGroup>
+    <!-- Debugger workaround https://github.com/dotnet/maui-samples/blob/8aa6b8780b12e97b157514c3bdc54bb4a13001cd/HelloMacCatalyst/HelloMacCatalyst.csproj#L7 -->
+    <!-- <MtouchExtraArgs Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'maccatalyst'">$(MtouchExtraArgs) -setenv:MONO_THREADS_SUSPEND=preemptive</MtouchExtraArgs> -->
+    <!-- Required for C# Hot Reload -->
+    <UseInterpreter Condition="'$(Configuration)' == 'Debug' and $([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) != 'maccatalyst'">True</UseInterpreter>
+    <IsUnoHead>true</IsUnoHead>
+  </PropertyGroup>
 
-	<!--<ItemGroup>
-		<TrimmableAssembly Include="Microsoft.Graph" />
-		<TrimmableAssembly Include="CommunityToolkit.WinUI" />
-	</ItemGroup>-->
+  <ItemGroup>
+    <PackageReference Include="Uno.WinUI" />
+    <PackageReference Include="CommunityToolkit.Mvvm" />
+    <PackageReference Include="Uno.Extensions.Configuration" />
+    <PackageReference Include="Uno.Extensions.Http" />
+    <PackageReference Include="Uno.Extensions.Http.Refit" />
+    <PackageReference Include="Uno.Extensions.Logging.WinUI" />
+    <PackageReference Include="Uno.Extensions.Serialization.Http" />
+    <PackageReference Include="Uno.Extensions.Serialization.Refit" />
+    <PackageReference Include="Uno.Toolkit.WinUI" />
+    <PackageReference Include="Uno.Extensions.Hosting.WinUI" />
+    <PackageReference Include="Uno.Extensions.Localization.WinUI" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" />
+    <PackageReference Include="Uno.Extensions.Logging.OSLog" />
+    <PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" />
+    <PackageReference Include="Uno.WinUI.DevServer" Condition="'$(Configuration)'=='Debug'" />
 
-	<ItemGroup>
-		<PackageReference Include="CommunityToolkit.Labs.WinUI.SettingsControls" />
-		<PackageReference Include="CommunityToolkit.WinUI.Controls.Segmented" />
-		<PackageReference Include="CommunityToolkit.WinUI.Controls.SettingsControls" />
-		<PackageReference Include="CommunityToolkit.WinUI.Converters" />
-		<PackageReference Include="CommunityToolkit.WinUI.Animations" />
-		<PackageReference Include="CommunityToolkit.WinUI.Triggers" />
-		<PackageReference Include="LiveChartsCore.SkiaSharpView.Uno.WinUI" />
-		<PackageReference Include="Microsoft.AppCenter" />
-		<PackageReference Include="Microsoft.AppCenter.Analytics" />
-		<PackageReference Include="Microsoft.AppCenter.Crashes" />
-		<PackageReference Include="Uno.WinUI.MSAL" />
-		<PackageReference Include="Uno.WinUI" />
-		<PackageReference Include="CommunityToolkit.Mvvm" />
-		<PackageReference Include="Uno.Extensions.Configuration" />
-		<PackageReference Include="Uno.Extensions.Http" />
-		<PackageReference Include="Uno.Extensions.Http.Refit" />
-		<PackageReference Include="Uno.Extensions.Logging.WinUI" />
-		<PackageReference Include="Uno.Extensions.Serialization.Http" />
-		<PackageReference Include="Uno.Extensions.Serialization.Refit" />
-		<PackageReference Include="Uno.Extensions.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Extensions.Authentication.WinUI" />
-		<PackageReference Include="Microsoft.Identity.Client" />
-		<PackageReference Include="Uno.Extensions.Hosting.WinUI" />
-		<PackageReference Include="Uno.Extensions.Localization.WinUI" />
-		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Extensions.Navigation.WinUI" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" />
-		<PackageReference Include="Uno.Extensions.Logging.OSLog" />
-		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" />
-		<PackageReference Include="Uno.WinUI.Lottie" />
-		<PackageReference Include="Uno.WinUI.RemoteControl" Condition="'$(Configuration)'=='Debug'" />
-		<PackageReference Include="Microsoft.Graph" />
-	</ItemGroup>
-	<Choose>
-		<When Condition="$(IsAndroid)">
-			<ItemGroup>
-				<PackageReference Include="Xamarin.Google.Android.Material" />
-				<PackageReference Include="Uno.UniversalImageLoader" />
-			</ItemGroup>
-			<ItemGroup>
-				<AndroidEnvironment Include="Android/environment.conf" />
-			</ItemGroup>
-		</When>
-		<When Condition="$(IsIOS)">
-			<PropertyGroup>
-				<MtouchExtraArgs>$(MtouchExtraArgs) --setenv=MONO_GC_PARAMS=soft-heap-limit=512m,nursery-size=64m,evacuation-threshold=66,major=marksweep,concurrent-sweep</MtouchExtraArgs>
-				<!-- See https://github.com/unoplatform/uno/issues/9430 for more details. -->
-				<MtouchExtraArgs>$(MtouchExtraArgs) --registrar:static</MtouchExtraArgs>
-				<RuntimeIdentifier Condition="'$(RuntimeIdentifier)'==''">iossimulator-x64</RuntimeIdentifier>
-			</PropertyGroup>
-			<PropertyGroup Condition="'$(Configuration)'=='Release'">
-				<!-- https://github.com/xamarin/xamarin-macios/issues/14812 -->
-				<MtouchExtraArgs>$(MtouchExtraArgs) --marshal-objectivec-exceptions:disable</MtouchExtraArgs>
-			</PropertyGroup>
-		</When>
-		<When Condition="$(IsMacCatalyst)">
-			<PropertyGroup>
-				<!-- Configure the GC -->
-				<MtouchExtraArgs>$(MtouchExtraArgs) --setenv=MONO_GC_PARAMS=soft-heap-limit=512m,nursery-size=64m,evacuation-threshold=66,major=marksweep,concurrent-sweep</MtouchExtraArgs>
-				<!-- Required for unknown crash as of .NET 6 Mobile Preview 13 -->
-				<MtouchExtraArgs>$(MtouchExtraArgs) --registrar:static</MtouchExtraArgs>
-				<!-- Full globalization is required for Uno -->
-				<InvariantGlobalization>false</InvariantGlobalization>
-				<RuntimeIdentifier Condition="'$(RuntimeIdentifier)'==''">maccatalyst-x64</RuntimeIdentifier>
-			</PropertyGroup>
-			<PropertyGroup Condition="'$(Configuration)'=='Release'">
-				<!-- https://github.com/xamarin/xamarin-macios/issues/14812 -->
-				<MtouchExtraArgs>$(MtouchExtraArgs) --marshal-objectivec-exceptions:disable</MtouchExtraArgs>
-			</PropertyGroup>
-		</When>
-	</Choose>
-	<ItemGroup>
-		<ProjectReference Include="..\ProjectSBS\ProjectSBS.csproj" />
-		<ProjectReference Include="..\ProjectSBS.DataContracts\ProjectSBS.DataContracts.csproj" />
-	</ItemGroup>
-	<Import Project="..\ProjectSBS.Base\base.props" />
+    <PackageReference Include="CommunityToolkit.WinUI.Notifications" />
+    <PackageReference Include="CommunityToolkit.WinUI.Controls.Segmented" />
+    <PackageReference Include="CommunityToolkit.WinUI.Controls.SettingsControls" />
+    <PackageReference Include="CommunityToolkit.WinUI.Converters" />
+    <PackageReference Include="CommunityToolkit.WinUI.Animations" />
+    <PackageReference Include="CommunityToolkit.WinUI.Triggers" />
+    <PackageReference Include="LiveChartsCore.SkiaSharpView.Uno.WinUI" />
+    <PackageReference Include="Microsoft.AppCenter" />
+    <PackageReference Include="Microsoft.AppCenter.Analytics" />
+    <PackageReference Include="Microsoft.AppCenter.Crashes" />
+    <PackageReference Include="Microsoft.Graph" />
+    <PackageReference Include="Microsoft.Identity.Client" />
+    <PackageReference Include="Uno.WinUI.MSAL" />
+  </ItemGroup>
+  <Choose>
+    <When Condition="$(IsAndroid)">
+      <ItemGroup>
+
+        <PackageReference Include="Xamarin.Google.Android.Material" />
+        <PackageReference Include="Uno.UniversalImageLoader" />
+      </ItemGroup>
+      <ItemGroup>
+        <AndroidEnvironment Include="Android/environment.conf" />
+      </ItemGroup>
+    </When>
+    <When Condition="$(IsIOS)">
+      <PropertyGroup>
+        <MtouchExtraArgs>$(MtouchExtraArgs) --setenv=MONO_GC_PARAMS=soft-heap-limit=512m,nursery-size=64m,evacuation-threshold=66,major=marksweep,concurrent-sweep</MtouchExtraArgs>
+        <!-- See https://github.com/unoplatform/uno/issues/9430 for more details. -->
+        <MtouchExtraArgs>$(MtouchExtraArgs) --registrar:static</MtouchExtraArgs>
+        <RuntimeIdentifier Condition="'$(RuntimeIdentifier)'==''">iossimulator-x64</RuntimeIdentifier>
+      </PropertyGroup>
+      <PropertyGroup Condition="'$(Configuration)'=='Release'">
+        <!-- https://github.com/xamarin/xamarin-macios/issues/14812 -->
+        <MtouchExtraArgs>$(MtouchExtraArgs) --marshal-objectivec-exceptions:disable</MtouchExtraArgs>
+      </PropertyGroup>
+    </When>
+    <When Condition="$(IsMacCatalyst)">
+      <PropertyGroup>
+        <!-- Configure the GC -->
+        <MtouchExtraArgs>$(MtouchExtraArgs) --setenv=MONO_GC_PARAMS=soft-heap-limit=512m,nursery-size=64m,evacuation-threshold=66,major=marksweep,concurrent-sweep</MtouchExtraArgs>
+        <!-- Required for unknown crash as of .NET 6 Mobile Preview 13 -->
+        <MtouchExtraArgs>$(MtouchExtraArgs) --registrar:static</MtouchExtraArgs>
+        <!-- Full globalization is required for Uno -->
+        <InvariantGlobalization>false</InvariantGlobalization>
+        <RuntimeIdentifier Condition="'$(RuntimeIdentifier)'==''">maccatalyst-x64</RuntimeIdentifier>
+      </PropertyGroup>
+      <PropertyGroup Condition="'$(Configuration)'=='Release'">
+        <!-- https://github.com/xamarin/xamarin-macios/issues/14812 -->
+        <MtouchExtraArgs>$(MtouchExtraArgs) --marshal-objectivec-exceptions:disable</MtouchExtraArgs>
+      </PropertyGroup>
+    </When>
+  </Choose>
+  <ItemGroup>
+    <ProjectReference Include="..\ProjectSBS\ProjectSBS.csproj" />
+    <ProjectReference Include="..\ProjectSBS.DataContracts\ProjectSBS.DataContracts.csproj" />
+  </ItemGroup>
+  <Import Project="..\ProjectSBS.Base\base.props" />
 </Project>
