@@ -1,104 +1,100 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
-	<PropertyGroup>
-		<OutputType>Exe</OutputType>
-		<TargetFramework>net7.0</TargetFramework>
-		<NoWarn>$(NoWarn);NU1504;NU1505;NU1701</NoWarn>
-		<!-- Disabled due to issue with Central Package Management with implicit using -->
-		<ImplicitUsings>disable</ImplicitUsings>
-		<WasmPWAManifestFile>manifest.webmanifest</WasmPWAManifestFile>
-		<AppDesignerFolder>Properties</AppDesignerFolder>
-		<!--
-			Supports Deep Linking Routes
-			https://aka.platform.uno/wasm-deeplink
-		-->
-		<WasmShellWebAppBasePath>/</WasmShellWebAppBasePath>
-	</PropertyGroup>
-	<PropertyGroup Condition="'$(Configuration)'=='Debug'">
-		<MonoRuntimeDebuggerEnabled>true</MonoRuntimeDebuggerEnabled>
-		<DefineConstants>$(DefineConstants);TRACE;DEBUG</DefineConstants>
-		<DebugType>portable</DebugType>
-		<DebugSymbols>true</DebugSymbols>
-		<!--
-			IL Linking is disabled in Debug configuration.
-			When building in Release, see https://platform.uno/docs/articles/features/using-il-linker-WebAssembly.html
-		-->
-		<WasmShellILLinkerEnabled>false</WasmShellILLinkerEnabled>
-	</PropertyGroup>
-	<PropertyGroup Condition="'$(Configuration)'=='Release'">
-		<!-- XAML Resource trimming https://aka.platform.uno/xaml-trimming -->
-		<!--<UnoXamlResourcesTrimming>true</UnoXamlResourcesTrimming>-->
-		<!-- Improve performance with AOT builds https://aka.platform.uno/wasm-aot -->
-		<!-- <WasmShellMonoRuntimeExecutionMode>InterpreterAndAOT</WasmShellMonoRuntimeExecutionMode> -->
-		<!-- Temporarily uncomment to generate an AOT profile https://aka.platform.uno/wasm-aot-profile -->
-		<!-- <WasmShellGenerateAOTProfile>true</WasmShellGenerateAOTProfile> -->
-	</PropertyGroup>
-	<PropertyGroup Condition="'$(Configuration)'=='Debug' or '$(IsUiAutomationMappingEnabled)'=='True'">
-		<IsUiAutomationMappingEnabled>True</IsUiAutomationMappingEnabled>
-		<DefineConstants>$(DefineConstants);USE_UITESTS</DefineConstants>
-	</PropertyGroup>
-	<ItemGroup>
-		<Content Include="manifest.webmanifest" />
-	</ItemGroup>
-	<ItemGroup>
-		<EmbeddedResource Include="WasmCSS\Fonts.css" />
-		<EmbeddedResource Include="WasmScripts\AppManifest.js" />
-	</ItemGroup>
-	<ItemGroup>
-		<LinkerDescriptor Include="LinkerConfig.xml" />
-	</ItemGroup>
-	<ItemGroup>
-		<!--
-		This item group is required by the project template because of the
-		new SDK-Style project, otherwise some files are not added automatically.
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <NoWarn>$(NoWarn);NU1504;NU1505;NU1701</NoWarn>
+    <!-- Disabled due to issue with Central Package Management with implicit using -->
+    <ImplicitUsings>disable</ImplicitUsings>
+    <WasmPWAManifestFile>manifest.webmanifest</WasmPWAManifestFile>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <!--
+      Supports Deep Linking Routes
+      https://aka.platform.uno/wasm-deeplink
+    -->
+    <WasmShellWebAppBasePath>/</WasmShellWebAppBasePath>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)'=='Debug'">
+    <MonoRuntimeDebuggerEnabled>true</MonoRuntimeDebuggerEnabled>
+    <DefineConstants>$(DefineConstants);TRACE;DEBUG</DefineConstants>
+    <DebugType>portable</DebugType>
+    <DebugSymbols>true</DebugSymbols>
+    <!--
+      IL Linking is disabled in Debug configuration.
+      When building in Release, see https://platform.uno/docs/articles/features/using-il-linker-WebAssembly.html
+    -->
+    <WasmShellILLinkerEnabled>false</WasmShellILLinkerEnabled>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)'=='Release'">
+    <!-- XAML Resource trimming https://aka.platform.uno/xaml-trimming -->
+    <!--<UnoXamlResourcesTrimming>true</UnoXamlResourcesTrimming>-->
+    <!-- Improve performance with AOT builds https://aka.platform.uno/wasm-aot -->
+    <!-- <WasmShellMonoRuntimeExecutionMode>InterpreterAndAOT</WasmShellMonoRuntimeExecutionMode> -->
+    <!-- Temporarily uncomment to generate an AOT profile https://aka.platform.uno/wasm-aot-profile -->
+    <!-- <WasmShellGenerateAOTProfile>true</WasmShellGenerateAOTProfile> -->
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)'=='Debug' or '$(IsUiAutomationMappingEnabled)'=='True'">
+    <IsUiAutomationMappingEnabled>True</IsUiAutomationMappingEnabled>
+    <DefineConstants>$(DefineConstants);USE_UITESTS</DefineConstants>
+  </PropertyGroup>
+  <ItemGroup>
+    <Content Include="manifest.webmanifest" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="WasmCSS\Fonts.css" />
+    <EmbeddedResource Include="WasmScripts\AppManifest.js" />
+  </ItemGroup>
+  <ItemGroup>
+    <LinkerDescriptor Include="LinkerConfig.xml" />
+  </ItemGroup>
+  <ItemGroup>
+    <!--
+    This item group is required by the project template because of the
+    new SDK-Style project, otherwise some files are not added automatically.
 
-		You can safely remove this ItemGroup completely.
-		-->
-		<None Include="Program.cs" />
-		<None Include="LinkerConfig.xml" />
-		<None Include="wwwroot\web.config" />
-	</ItemGroup>
+    You can safely remove this ItemGroup completely.
+    -->
+    <None Include="Program.cs" />
+    <None Include="LinkerConfig.xml" />
+    <None Include="wwwroot\web.config" />
+  </ItemGroup>
 
-	<ItemGroup>
-		<PackageReference Include="CommunityToolkit.Labs.WinUI.SettingsControls" />
-		<PackageReference Include="CommunityToolkit.WinUI.Controls.Segmented" />
-		<PackageReference Include="CommunityToolkit.WinUI.Controls.SettingsControls" />
-		<PackageReference Include="CommunityToolkit.WinUI.Converters" />
-		<PackageReference Include="CommunityToolkit.WinUI.Animations" />
-		<PackageReference Include="CommunityToolkit.WinUI.Triggers" />
-		<PackageReference Include="LiveChartsCore.SkiaSharpView.Uno.WinUI" />
-		<PackageReference Include="Microsoft.AppCenter" />
-		<PackageReference Include="Microsoft.AppCenter.Analytics" />
-		<PackageReference Include="Microsoft.AppCenter.Crashes" />
-		<PackageReference Include="Microsoft.Windows.Compatibility" />
-		<PackageReference Include="Uno.Wasm.Bootstrap" />
-		<PackageReference Include="Uno.Wasm.Bootstrap.DevServer" />
-		<PackageReference Include="Uno.WinUI.MSAL" />
-		<PackageReference Include="Uno.WinUI.Lottie" />
-		<PackageReference Include="Uno.WinUI.WebAssembly" />
-		<PackageReference Include="Uno.WinUI.RemoteControl" Condition="'$(Configuration)'=='Debug'" />
-		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" />
-		<PackageReference Include="CommunityToolkit.Mvvm" />
-		<PackageReference Include="Uno.Extensions.Configuration" />
-		<PackageReference Include="Uno.Extensions.Http" />
-		<PackageReference Include="Uno.Extensions.Http.Refit" />
-		<PackageReference Include="Uno.Extensions.Logging.WebAssembly.Console" />
-		<PackageReference Include="Uno.Extensions.Logging.WinUI" />
-		<PackageReference Include="Uno.Extensions.Serialization.Http" />
-		<PackageReference Include="Uno.Extensions.Serialization.Refit" />
-		<PackageReference Include="Uno.Extensions.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Extensions.Authentication.WinUI" />
-		<PackageReference Include="Microsoft.Identity.Client" />
-		<PackageReference Include="Uno.Extensions.Hosting.WinUI" />
-		<PackageReference Include="Uno.Extensions.Localization.WinUI" />
-		<PackageReference Include="Uno.Extensions.Navigation.Toolkit.WinUI" />
-		<PackageReference Include="Uno.Extensions.Navigation.WinUI" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" />
-		<PackageReference Include="Microsoft.Graph" />
-	</ItemGroup>
-	<ItemGroup>
-		<ProjectReference Include="..\ProjectSBS\ProjectSBS.csproj" />
-		<ProjectReference Include="..\ProjectSBS.DataContracts\ProjectSBS.DataContracts.csproj" />
-	</ItemGroup>
-	<Import Project="..\ProjectSBS.Base\base.props" />
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Windows.Compatibility" />
+    <PackageReference Include="Uno.Wasm.Bootstrap" />
+    <PackageReference Include="Uno.Wasm.Bootstrap.DevServer" />
+    <PackageReference Include="Uno.WinUI.WebAssembly" />
+    <PackageReference Include="Uno.WinUI.DevServer" Condition="'$(Configuration)'=='Debug'" />
+    <PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" />
+    <PackageReference Include="CommunityToolkit.Mvvm" />
+    <PackageReference Include="Uno.Extensions.Configuration" />
+    <PackageReference Include="Uno.Extensions.Http" />
+    <PackageReference Include="Uno.Extensions.Http.Refit" />
+    <PackageReference Include="Uno.Extensions.Logging.WebAssembly.Console" />
+    <PackageReference Include="Uno.Extensions.Logging.WinUI" />
+    <PackageReference Include="Uno.Extensions.Serialization.Http" />
+    <PackageReference Include="Uno.Extensions.Serialization.Refit" />
+    <PackageReference Include="Uno.Toolkit.WinUI" />
+    <PackageReference Include="Uno.Extensions.Hosting.WinUI" />
+    <PackageReference Include="Uno.Extensions.Localization.WinUI" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" />
+
+    <PackageReference Include="CommunityToolkit.WinUI.Notifications" />
+    <PackageReference Include="CommunityToolkit.WinUI.Controls.Segmented" />
+    <PackageReference Include="CommunityToolkit.WinUI.Controls.SettingsControls" />
+    <PackageReference Include="CommunityToolkit.WinUI.Converters" />
+    <PackageReference Include="CommunityToolkit.WinUI.Animations" />
+    <PackageReference Include="CommunityToolkit.WinUI.Triggers" />
+    <PackageReference Include="LiveChartsCore.SkiaSharpView.Uno.WinUI" />
+    <PackageReference Include="Microsoft.AppCenter" />
+    <PackageReference Include="Microsoft.AppCenter.Analytics" />
+    <PackageReference Include="Microsoft.AppCenter.Crashes" />
+    <PackageReference Include="Microsoft.Graph" />
+    <PackageReference Include="Microsoft.Identity.Client" />
+    <PackageReference Include="Uno.WinUI.MSAL" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\ProjectSBS\ProjectSBS.csproj" />
+    <ProjectReference Include="..\ProjectSBS.DataContracts\ProjectSBS.DataContracts.csproj" />
+  </ItemGroup>
+  <Import Project="..\ProjectSBS.Base\base.props" />
 </Project>
