@@ -1,3 +1,7 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
 using Android.App;
 using Android.Content;
 using Android.OS;
@@ -6,38 +10,33 @@ using Android.Views;
 using Android.Widget;
 using Com.Nostra13.Universalimageloader.Core;
 using Microsoft.UI.Xaml.Media;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
-namespace ProjectSBS.Droid
+namespace ProjectSBS.Droid;
+[global::Android.App.ApplicationAttribute(
+    Label = "@string/ApplicationName",
+    Icon = "@mipmap/icon",
+    LargeHeap = true,
+    HardwareAccelerated = true,
+    Theme = "@style/AppTheme"
+)]
+public class Application : Microsoft.UI.Xaml.NativeApplication
 {
-    [global::Android.App.ApplicationAttribute(
-        Label = "@string/ApplicationName",
-        Icon = "@mipmap/iconapp",
-        LargeHeap = true,
-        HardwareAccelerated = true,
-        Theme = "@style/AppTheme"
-    )]
-    public class Application : Microsoft.UI.Xaml.NativeApplication
+    public Application(IntPtr javaReference, JniHandleOwnership transfer)
+        : base(() => new AppHead(), javaReference, transfer)
     {
-        public Application(IntPtr javaReference, JniHandleOwnership transfer)
-            : base(() => new AppHead(), javaReference, transfer)
-        {
-            ConfigureUniversalImageLoader();
-        }
+        ConfigureUniversalImageLoader();
+    }
 
-        private static void ConfigureUniversalImageLoader()
-        {
-            // Create global configuration and initialize ImageLoader with this config
-            ImageLoaderConfiguration config = new ImageLoaderConfiguration
-                .Builder(Context)
-                .Build();
+    private static void ConfigureUniversalImageLoader()
+    {
+        // Create global configuration and initialize ImageLoader with this config
+        ImageLoaderConfiguration config = new ImageLoaderConfiguration
+            .Builder(Context)
+            .Build();
 
-            ImageLoader.Instance.Init(config);
+        ImageLoader.Instance.Init(config);
 
-            ImageSource.DefaultImageLoader = ImageLoader.Instance.LoadImageAsync;
-        }
+        ImageSource.DefaultImageLoader = ImageLoader.Instance.LoadImageAsync;
     }
 }
+
