@@ -3,21 +3,13 @@ using Android.Content.PM;
 using Android.OS;
 using Android.Views;
 using Android.Widget;
-using Microsoft.Identity.Client;
 
-namespace ProjectSBS.Droid
+namespace ProjectSBS.Droid;
+[Activity(
+    MainLauncher = true,
+    ConfigurationChanges = global::Uno.UI.ActivityHelper.AllConfigChanges,
+    WindowSoftInputMode = SoftInput.AdjustNothing | SoftInput.StateHidden
+)]
+public class MainActivity : Microsoft.UI.Xaml.ApplicationActivity
 {
-    [Activity(
-        MainLauncher = true,
-        ConfigurationChanges = global::Uno.UI.ActivityHelper.AllConfigChanges,
-        WindowSoftInputMode = SoftInput.AdjustNothing | SoftInput.StateHidden
-    )]
-    public class MainActivity : Microsoft.UI.Xaml.ApplicationActivity
-    {
-        protected override void OnActivityResult(int requestCode, Result resultCode, Android.Content.Intent data)
-        {
-            base.OnActivityResult(requestCode, resultCode, data);
-            AuthenticationContinuationHelper.SetAuthenticationContinuationEventArgs(requestCode, resultCode, data);
-        }
-    }
 }
