@@ -1,15 +1,13 @@
 using UIKit;
 
-namespace ProjectSBS.iOS
+namespace ProjectSBS.iOS;
+public class EntryPoint
 {
-    public class EntryPoint
+    // This is the main entry point of the application.
+    public static void Main(string[] args)
     {
-        // This is the main entry point of the application.
-        public static void Main(string[] args)
-        {
-            // if you want to use a different Application Delegate class from "AppDelegate"
-            // you can specify it here.
-            UIApplication.Main(args, null, typeof(AppHead));
-        }
+        // if you want to use a different Application Delegate class from "AppDelegate"
+        // you can specify it here.
+        UIApplication.Main(args, null, typeof(AppHead));
     }
 }
