@@ -1,19 +1,18 @@
-namespace ProjectSBS.Tests
+namespace ProjectSBS.Tests;
+
+public class AppInfoTests
 {
-    public class AppInfoTests
+    [SetUp]
+    public void Setup()
     {
-        [SetUp]
-        public void Setup()
-        {
-        }
+    }
 
-        [Test]
-        public void AppInfoCreation()
-        {
-            var appInfo = new AppConfig { Environment = "Test" };
+    [Test]
+    public void AppInfoCreation()
+    {
+        var appInfo = new AppConfig { Environment = "Test" };
 
-            appInfo.Should().NotBeNull();
-            appInfo.Environment.Should().Be("Test");
-        }
+        appInfo.Should().NotBeNull();
+        appInfo.Environment.Should().Be("Test");
     }
 }
