@@ -1,12 +1,12 @@
-﻿namespace ProjectSBS.UITests
-{
-    public class Constants
-    {
-        public readonly static string WebAssemblyDefaultUri = "http://localhost:5000/";
-        public readonly static string iOSAppName = "com.morning4coffe.ProjectSBS";
-        public readonly static string AndroidAppName = "com.morning4coffe.ProjectSBS";
-        public readonly static string iOSDeviceNameOrId = "iPad Pro (12.9-inch) (3rd generation)";
+namespace ProjectSBS.UITests;
 
-        public readonly static Platform CurrentPlatform = Platform.Browser;
-    }
+public class Constants
+{
+    public readonly static string WebAssemblyDefaultUri = "http://localhost:5001/";
+    public readonly static string iOSAppName = "com.morning4coffe.ProjectSBS";
+    public readonly static string AndroidAppName = "com.morning4coffe.ProjectSBS";
+    public readonly static string iOSDeviceNameOrId = "iPad Pro (12.9-inch) (3rd generation)";
+
+    public readonly static Platform CurrentPlatform = Platform.Browser;
+    public readonly static Browser WebAssemblyBrowser = Browser.Chrome;
 }
