@@ -1,14 +1,13 @@
-namespace ProjectSBS.Wasm
+namespace ProjectSBS.Wasm;
+
+public class Program
 {
-    public class Program
+    private static App? _app;
+
+    public static int Main(string[] args)
     {
-        private static App? _app;
+        Microsoft.UI.Xaml.Application.Start(_ => _app = new AppHead());
 
-        public static int Main(string[] args)
-        {
-            Microsoft.UI.Xaml.Application.Start(_ => _app = new AppHead());
-
-            return 0;
-        }
+        return 0;
     }
 }
