@@ -1,4 +1,3 @@
-namespace ProjectSBS.Business.Models
-{
-    public record Entity(string Name);
-}
+namespace ProjectSBS.Business.Models;
+
+public record Entity(string Name);
