@@ -1,3 +1,3 @@
-﻿global using FluentAssertions;
+global using FluentAssertions;
 global using NUnit.Framework;
 global using ProjectSBS.Business.Models;
