@@ -1,4 +1,4 @@
-﻿global using NUnit.Framework;
+global using NUnit.Framework;
 global using Uno.UITest;
 global using Uno.UITest.Helpers.Queries;
 global using Uno.UITests.Helpers;
