@@ -1,20 +1,13 @@
-﻿namespace ProjectSBS.Business.Models;
+namespace ProjectSBS.Business.Models;
 
-public class ItemLog
+public class ItemLog(
+    string itemId,
+    DateOnly paymentDate,
+    decimal price,
+    string currencyId)
 {
-    public ItemLog( 
-        string itemId, 
-        DateOnly paymentDate, 
-        decimal price, 
-        string currencyId)
-    {
-        ItemId = itemId;
-        PaymentDate = paymentDate;
-        Price = price;
-        CurrencyId = currencyId;
-    }
-    public string ItemId { get; set; }
-    public DateOnly PaymentDate { get; set; }
-    public decimal Price { get; set; }
-    public string CurrencyId { get; set; }
+    public string ItemId { get; set; } = itemId;
+    public DateOnly PaymentDate { get; set; } = paymentDate;
+    public decimal Price { get; set; } = price;
+    public string CurrencyId { get; set; } = currencyId;
 }
