@@ -1,20 +1,13 @@
-﻿using System.Drawing;
+using System.Drawing;
 
 namespace ProjectSBS.Business.Models;
 
-public class Tag 
+public class Tag(
+    int id,
+    string name,
+    Color? color)
 {
-    public int Id { get; set; }
-    public string Name { get; set; }
-    public Color? Color { get; set; }
-
-    public Tag(
-        int id, 
-        string name, 
-        Color? color)
-    {
-        Id = id;
-        Name = name;
-        Color = color;
-    }
+    public int Id { get; set; } = id;
+    public string Name { get; set; } = name;
+    public Color? Color { get; set; } = color;
 }
