@@ -1,4 +1,4 @@
-﻿namespace ProjectSBS.Services.Navigation;
+namespace ProjectSBS.Services.Navigation;
 
 internal class NavigationService : INavigation
 {
@@ -11,15 +11,15 @@ internal class NavigationService : INavigation
 
     public NavigationService(IStringLocalizer localizer)
     {
-        Categories = new()
-        {
+        Categories =
+        [
             new(0, localizer["Home"], "\uE80F", typeof(HomePage)),
             new(1, localizer["Items"], "\uF0B2", typeof(HomePage)),
-            new(2, localizer["Archive"], "\uE7B8", typeof(HomePage)),
-            new(3, localizer["Stats"], "\uEAFC", typeof(HomePage)),
-            new(4, localizer["Dev"], "\uE98F", typeof(HomePage)/*, CategoryVisibility.Desktop*/),
+            new(2, localizer["Archive"], "\uE7B8", typeof(ArchivePage)),
+            //new(3, localizer["Stats"], "\uEAFC", typeof(HomePage)),
+            //new(4, localizer["Dev"], "\uE98F", typeof(HomePage)/*, CategoryVisibility.Desktop*/),
             new(5, localizer["Settings"], "\uE713", typeof(SettingsPage)/*, CategoryVisibility.Mobile*/),
-        };
+        ];
 
         SelectedCategory = Categories[0];
     }
